requirements.txt
streamlit
Pillow
python-dotenv
streamlit==1.40.0
google-generativeai